Stop cleaning of candidate directory if candidate is already scored

+++ src/main/java/nl/bioinf/cawarmerdam/compound_evolver/control/CompoundEvolver.java
@@ -612,7 +612,7 @@ public class CompoundEvolver {
                     File candidate_dir = Paths.get(pipelineOutputFilePath.toString(),
                             String.valueOf(candidate.getIdentifier())).toFile();
                     try {
-                        if (candidate_dir.exists())
+                        if (candidate_dir.exists() && candidate.canBeDeleted())
                             FileUtils.cleanDirectory(candidate_dir);
                     } catch (IOException e) {
                         if (debugPrint) {
